Move frame-related classes from track.model to domain.frame
This refactor relocates `DetectedFrame`, `TrackedFrame`, and related classes to a more appropriate module under `domain.frame`. All imports across the codebase were updated accordingly to reflect the new location, ensuring consistency and maintainability.
