+++ OTVision/track/model/tracking_interfaces.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from typing import Generic, Iterator, TypeVar
 
 from OTVision.domain.detection import TrackId
-from OTVision.track.model.frame import (
+from OTVision.domain.frame import (
     DetectedFrame,
     FinishedFrame,
     FrameNo,
